Deduplicate convention-specific test files (#608)
* Deduplicate convention-specific test files into single files

Merge 4 pairs of v1/legacy test files into single files, eliminating
~2600 lines of duplicated test code. Convention-specific alignment tests
are kept in separate classes (V1/Legacy) with autouse fixtures, while
shared tests run under the module-level v1 convention.

- test_typing_legacy.py -> merged into test_typing.py (parametrized)
- test_common_legacy.py -> merged into test_common.py (legacy align test)
- test_constraints_legacy.py -> merged into test_constraints.py (legacy alignment class)
- test_linear_expression_legacy.py -> merged into test_linear_expression.py (legacy alignment + join classes)


* Address PR review: consistency, dedup fixtures, missing test

- Add legacy_convention fixture to conftest.py; use it consistently
  instead of manual try/finally blocks (#1)
- Parametrize test_constant_with_extra_dims_broadcasts with convention
  fixture so it runs under both conventions (#2)
- Add missing test_quadratic_add_expr_join_inner to
  TestJoinParameterLegacy (#3)
- Extract shared fixtures into _CoordinateAlignmentFixtures and
  _ConstraintAlignmentFixtures mixin classes to eliminate fixture
  duplication between V1/Legacy alignment test classes (#4)


---------
